+++ b_twilio_sdk_layer_test/integration/app.py
@@ -9,12 +9,12 @@ Import main stack.
 
 ROOT_PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(ROOT_PATH)
-from b_twilio_sdk_layer_test.testing_infrastructure import TestingInfrastructure
+from b_twilio_sdk_layer_test.integration.infrastructure import Infrastructure
 
 """
 Create CDK app.
 """
 
 app = App()
-TestingInfrastructure(app)
+Infrastructure(app)
 app.synth()